This is a massive intertwined commit that supports flats as walls and walls as flats.
Lots of ripping out of old arrays and integrating new ones. Both textures and flats have separate lookup arrays that point back to the master tables. Will be easier this way to keep vanilla compatibility.

Animation specails in the playsim have been updated as a result. This needs to move out to the renderer to be quite honest.

+++ src/doom/r_raster.cpp
@@ -191,7 +191,7 @@ INLINE void R_Prepare( int32_t y, planecontext_t* planecontext )
 		}
 		else
 		{
-			int32_t lightindex = M_CLAMP( ( planecontext->raster[ y ].distance >> RENDLIGHTZSHIFT ), 0, ( MAXLIGHTZ - 1 ) );
+			int32_t lightindex = (int32_t)M_CLAMP( ( planecontext->raster[ y ].distance >> RENDLIGHTZSHIFT ), 0, ( MAXLIGHTZ - 1 ) );
 			lightindex = zlightindex[ planecontext->planezlightindex ][ lightindex ];
 			planecontext->raster[ y ].sourceoffset = lightindex * 4096;
 		}
